Rename hermite0 -> hermite
and keep old naming for compatibility
