+++ packages/amuse-hermite/src/amuse/community/hermite
@@ -1,0 +1,1 @@
+../../../../../src/amuse/community/hermite
