+++ src/amuse/community/hermite/vec3.h
