--- src/amuse/community/hermite0/__init__.py
@@ -1,1 +1,0 @@
-
